+++ src/main/java/com/mysawit/pembayaran/service/strategy/WageCalculationStrategy.java
@@ -2,7 +2,9 @@ package com.mysawit.pembayaran.service.strategy;
 
 import com.mysawit.pembayaran.model.enums.UserRole;
 
+import java.math.BigDecimal;
+
 public interface WageCalculationStrategy {
-    double calculate(double wagePerKg, double kilogram);
+    BigDecimal calculate(BigDecimal wagePerKg, BigDecimal kilogram);
     UserRole getSupportedRole();
 }
